README.md: sample configuration toggle_telescope fix
ipairs doesn't iterate over all the keys in table. It stops when it sees a nil. If we remove any file in the harpoon:list() we set it to nil. The harpoon's quick_menu ui shows blank for the removed file, however the telescope ui only lists file till it encounters a nil.

This same issue might be causing the marked files to disappear on reopening nvim?